release: v1.8.1

+++ taosx-core/src/plugins/runners/mssql/appender/mod.rs
@@ -364,7 +364,7 @@ pub fn to_record_batches(
                         // convert to seconds and nanoseconds(since 1900-01-01 00:00:00, and seconds are actually 1/300 seconds)
                         let secs = (days as u32 - 25567) * 24 * 60 * 60 + secs / 300;
                         // convert to datetime with timezone
-                        let datetime = DateTime::from_timestamp(secs as i64, 0 as u32).unwrap();
+                        let datetime = DateTime::from_timestamp(secs as i64, 0_u32).unwrap();
 
                         builders[col_cidx]
                             .as_any_mut()
@@ -387,7 +387,7 @@ pub fn to_record_batches(
                         // convert to seconds and nanoseconds(since 1900-01-01 00:00:00, and seconds are actually minutes)
                         let secs = (days as i64 - 25567) * 24 * 60 * 60 + (secs as i64) * 60;
                         // convert to datetime with timezone
-                        let datetime = DateTime::from_timestamp(secs, 0 as u32).unwrap();
+                        let datetime = DateTime::from_timestamp(secs, 0_u32).unwrap();
 
                         builders[col_cidx]
                             .as_any_mut()
@@ -495,7 +495,7 @@ pub fn to_record_batches(
                             .as_any_mut()
                             .downcast_mut::<array::TimestampNanosecondBuilder>()
                             .unwrap()
-                            .append_value(datetime.timestamp_nanos_opt().unwrap() as i64);
+                            .append_value(datetime.timestamp_nanos_opt().unwrap());
                     }
                 },
             }
@@ -647,6 +647,7 @@ mod tests {
         }
     }
 
+    #[ignore]
     #[tokio::test]
     async fn test_to_schema() {
         // prepare data
@@ -682,6 +683,7 @@ mod tests {
         let _ = test_clear_data().await;
     }
 
+    #[ignore]
     #[tokio::test]
     async fn test_to_record_batch() {
         // prepare data
@@ -718,6 +720,7 @@ mod tests {
         let _ = test_clear_data().await;
     }
 
+    #[ignore]
     #[tokio::test]
     async fn test_to_record_batches() {
         // prepare data
